requirements.txt
django-compressor==2.0
django-crispy-forms==1.7.2
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
